+++ packages/client/src/api/controlPlaneSchemas.ts
@@ -109,6 +109,10 @@ export type DatabaseMetadata = {
    */
   createdAt: DateTime;
   /**
+   * @x-internal true
+   */
+  newMigrations?: boolean;
+  /**
    * Metadata about the database for display in Xata user interfaces
    */
   ui?: {
@@ -144,11 +148,15 @@ export type BranchMetadata = {
 };
 
 /**
+ * @maxLength 255
+ * @minLength 1
  * @pattern [a-zA-Z0-9_\-~]+
  */
 export type BranchName = string;
 
 /**
+ * @maxLength 255
+ * @minLength 1
  * @pattern [a-zA-Z0-9_\-~]+
  */
 export type DBName = string;