+++ packages/client/src/schema/record.ts
@@ -128,15 +128,21 @@ export function isXataRecord(x: any): x is XataRecord & Record<string, unknown> 
   return isIdentifiable(x) && isObject(metadata) && typeof metadata.version === 'number';
 }
 
+type EditableDataFields<T> = T extends XataRecord
+  ? { id: string } | string
+  : NonNullable<T> extends XataRecord
+  ? { id: string } | string | null | undefined
+  : T extends Date
+  ? string | Date
+  : NonNullable<T> extends Date
+  ? string | Date | null | undefined
+  : T;
+
 export type EditableData<O extends XataRecord> = Identifiable &
   Partial<
     Omit<
       {
-        [K in keyof O]: O[K] extends XataRecord
-          ? { id: string } | string
-          : NonNullable<O[K]> extends XataRecord
-          ? { id: string } | string | null | undefined
-          : O[K];
+        [K in keyof O]: EditableDataFields<O[K]>;
       },
       keyof XataRecord
     >